fix bug login
