+++ data_buku.txt
@@ -1,0 +1,3 @@
+Your Name,               Makoto Shinkai
+Weathering With You,     Makoto Shinkai
+Re:ZERO,                 Tapei Nagatsuki
