--- data_peminjaman.txt
@@ -1,1 +1,0 @@
-Your Name,		Nov 29 2022
